chore: move db utils to app
