+++ apps/wallet/src/utils/storage/index.ts
@@ -1,7 +1,2 @@
-import WebSQLite from 'react-native-quick-websql';
-import { create } from '@walless/store';
-import { SQLiteAdapterFactory } from 'pouchdb-adapter-react-native-sqlite';
-
-const SQLiteAdapter = SQLiteAdapterFactory(WebSQLite);
-
-export const storage = create('engine', SQLiteAdapter);
+export * from './collectibles';
+export * from './db';
